@@ -5,7 +5,7 @@ authors = [
     { name="Ben Johnson", email="benjamin.johnson@cfa.harvard.edu" },
 ]
 readme = "README.md"
-requires-python = ">=3.7"
+requires-python = ">=3.9"
 license = { text = "MIT License" }
 classifiers = [
     "Development Status :: 5 - Production/Stable",
@@ -20,7 +20,6 @@ dependencies = ["numpy", "h5py"]
 
 [project.optional-dependencies]
 test = ["pytest", "pytest-xdist"]
-
 
 [tool.setuptools]
 packages = ["prospect",
